platform-control-plane/requirements.txt
# Platform Control Plane Dependencies

# Core MCP
mcp>=0.1.0

# AWS Provider
boto3>=1.26.0
botocore>=1.29.0

# Azure Provider  
azure-mgmt-compute>=30.0.0
azure-mgmt-network>=23.0.0
azure-mgmt-resource>=22.0.0
azure-identity>=1.14.0

# GCP Provider
google-cloud-compute>=1.14.0
google-cloud-storage>=2.10.0
google-auth>=2.23.0

# SentinelOne
requests>=2.31.0
aiohttp>=3.9.0

# Cisco
netmiko>=4.2.0
paramiko>=3.3.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
asyncio>=3.4.3
aiofiles>=23.2.0
tenacity>=8.2.0  # For retries
cachetools>=5.3.0  # For caching

# Security
cryptography>=41.0.0

# Logging and Monitoring
structlog>=23.2.0
prometheus-client>=0.18.0  # For metrics export

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
mypy>=1.7.0
